+++ types/index.d.ts
@@ -100,6 +100,9 @@ type PluginEvent = keyof HLJSPlugin;
 type HLJSPlugin = {
     'after:highlight'?: (result: HighlightResult) => void,
     'before:highlight'?: (context: BeforeHighlightContext) => void,
+    'after:highlightElement'?: (data: { block: Element, result: HighlightResult, text: string}) => void,
+    'before:highlightElement'?: (data: { block: Element, language: string}) => void,
+    // TODO: Old API, remove with v12
     'after:highlightBlock'?: (data: { block: Element, result: HighlightResult, text: string}) => void,
     'before:highlightBlock'?: (data: { block: Element, language: string}) => void,
 }